+++ src/basic/App.tsx
@@ -1,9 +1,11 @@
 import { useState, useEffect } from 'react';
 
-import type { CartItem, Coupon, Product } from '../types';
-import { AdminPage } from './components/AdminPage';
-import { CartPage } from './components/CartPage';
+import type { CartItem, Coupon, Notification as TNotification, ProductWithUI } from '../types';
 import Header from './components/Header';
+import Notification from './components/Notification';
+import { initialCoupons, initialProducts } from './constants';
+import { AdminPage } from './pages/AdminPage';
+import { CartPage } from './pages/CartPage';
 
 // TODO: 메인 App 컴포넌트
 // 힌트:
@@ -11,67 +13,6 @@ import Header from './components/Header';
 // 2. 네비게이션 바에 모드 전환 버튼 포함
 // 3. 조건부 렌더링으로 CartPage 또는 AdminPage 표시
 // 4. 상태 관리는 각 페이지 컴포넌트에서 처리 (App은 라우팅만 담당)
-
-interface ProductWithUI extends Product {
-  description?: string;
-  isRecommended?: boolean;
-}
-
-interface Notification {
-  id: string;
-  message: string;
-  type: 'error' | 'success' | 'warning';
-}
-
-// 초기 데이터
-const initialProducts: ProductWithUI[] = [
-  {
-    id: 'p1',
-    name: '상품1',
-    price: 10000,
-    stock: 20,
-    discounts: [
-      { quantity: 10, rate: 0.1 },
-      { quantity: 20, rate: 0.2 },
-    ],
-    description: '최고급 품질의 프리미엄 상품입니다.',
-  },
-  {
-    id: 'p2',
-    name: '상품2',
-    price: 20000,
-    stock: 20,
-    discounts: [{ quantity: 10, rate: 0.15 }],
-    description: '다양한 기능을 갖춘 실용적인 상품입니다.',
-    isRecommended: true,
-  },
-  {
-    id: 'p3',
-    name: '상품3',
-    price: 30000,
-    stock: 20,
-    discounts: [
-      { quantity: 10, rate: 0.2 },
-      { quantity: 30, rate: 0.25 },
-    ],
-    description: '대용량과 고성능을 자랑하는 상품입니다.',
-  },
-];
-
-const initialCoupons: Coupon[] = [
-  {
-    name: '5000원 할인',
-    code: 'AMOUNT5000',
-    discountType: 'amount',
-    discountValue: 5000,
-  },
-  {
-    name: '10% 할인',
-    code: 'PERCENT10',
-    discountType: 'percentage',
-    discountValue: 10,
-  },
-];
 
 const App = () => {
   const [products, setProducts] = useState<ProductWithUI[]>(() => {
@@ -111,7 +52,7 @@ const App = () => {
   });
 
   const [isAdmin, setIsAdmin] = useState(false);
-  const [notifications, setNotifications] = useState<Notification[]>([]);
+  const [notifications, setNotifications] = useState<TNotification[]>([]);
   const [debouncedSearchTerm, setDebouncedSearchTerm] = useState('');
 
   const [totalItemCount, setTotalItemCount] = useState(0);
@@ -140,35 +81,7 @@ const App = () => {
   return (
     <div className='min-h-screen bg-gray-50'>
       {notifications.length > 0 && (
-        <div className='fixed top-20 right-4 z-50 space-y-2 max-w-sm'>
-          {notifications.map((notif) => (
-            <div
-              key={notif.id}
-              className={`p-4 rounded-md shadow-md text-white flex justify-between items-center ${
-                notif.type === 'error'
-                  ? 'bg-red-600'
-                  : notif.type === 'warning'
-                    ? 'bg-yellow-600'
-                    : 'bg-green-600'
-              }`}
-            >
-              <span className='mr-2'>{notif.message}</span>
-              <button
-                onClick={() => setNotifications((prev) => prev.filter((n) => n.id !== notif.id))}
-                className='text-white hover:text-gray-200'
-              >
-                <svg className='w-4 h-4' fill='none' stroke='currentColor' viewBox='0 0 24 24'>
-                  <path
-                    strokeLinecap='round'
-                    strokeLinejoin='round'
-                    strokeWidth={2}
-                    d='M6 18L18 6M6 6l12 12'
-                  />
-                </svg>
-              </button>
-            </div>
-          ))}
-        </div>
+        <Notification notifications={notifications} setNotifications={setNotifications} />
       )}
 
       <Header
