+++ src/basic/constants/index.ts
@@ -1,8 +1,59 @@
-// TODO: 초기 데이터 상수
+import { ProductWithUI, Coupon } from '../../types';
+
 // 정의할 상수들:
 // - initialProducts: 초기 상품 목록 (상품1, 상품2, 상품3 + 설명 필드 포함)
 // - initialCoupons: 초기 쿠폰 목록 (5000원 할인, 10% 할인)
 //
 // 참고: origin/App.tsx의 초기 데이터 구조를 참조
 
-// TODO: 구현
+// 초기 데이터
+const initialProducts: ProductWithUI[] = [
+  {
+    id: 'p1',
+    name: '상품1',
+    price: 10000,
+    stock: 20,
+    discounts: [
+      { quantity: 10, rate: 0.1 },
+      { quantity: 20, rate: 0.2 },
+    ],
+    description: '최고급 품질의 프리미엄 상품입니다.',
+  },
+  {
+    id: 'p2',
+    name: '상품2',
+    price: 20000,
+    stock: 20,
+    discounts: [{ quantity: 10, rate: 0.15 }],
+    description: '다양한 기능을 갖춘 실용적인 상품입니다.',
+    isRecommended: true,
+  },
+  {
+    id: 'p3',
+    name: '상품3',
+    price: 30000,
+    stock: 20,
+    discounts: [
+      { quantity: 10, rate: 0.2 },
+      { quantity: 30, rate: 0.25 },
+    ],
+    description: '대용량과 고성능을 자랑하는 상품입니다.',
+  },
+];
+
+const initialCoupons: Coupon[] = [
+  {
+    name: '5000원 할인',
+    code: 'AMOUNT5000',
+    discountType: 'amount',
+    discountValue: 5000,
+  },
+  {
+    name: '10% 할인',
+    code: 'PERCENT10',
+    discountType: 'percentage',
+    discountValue: 10,
+  },
+];
+
+export { initialProducts, initialCoupons };
